Extractions and refactors
